backend/requirements.txt
fastapi
python-dotenv
langchain-community
langchain-openai
python-multipart
unstructured
motor==3.2.0
uvicorn==0.22.0
pydantic[email]==2.0.3
python-decouple==3.8
duckduckgo-search 
langchain-experimental
wikipedia
litellm
langchain-mongodb
pdfplumber
Markdown
pdfminer
docx2txt
networkx
langchain 
pymongo 
tiktoken
faiss-cpu
llmx 
openai 
lida
langchain
langchain_community